This commit was manufactured by cvs2svn to create tag 'FSL6-0-0'.
Sprout from master 2017-07-05 14:45:27 UTC Saad Jbabdi <saad@fmrib.ox.ac.uk> 'small bug fix'
Cherrypick from master 2013-07-09 11:06:37 UTC Moises Fernandez <moisesf@fmrib.ox.ac.uk> 'Change FSLDEVDIR path with FSLDIR':
    CUDA/bedpostx_multigpu_LSF
Cherrypick from newnew 2017-01-11 16:14:05 UTC FSL Automated Build Engineer <cowboy@fmrib.ox.ac.uk> 'Merged from 510 stable':
    Makefile
    basgen.cc
    dtifit.cc
    dtigen.cc
    make_dyadic_vectors.cc
    ptx_meshmask.cc
    streamlines.cc
    vecreg.cc
Delete:
    CUDA/bedpostx_multigpu_PBS.sh
    RLR_MAP_routines.cpp
    RLR_MAP_routines.h
    doc/fdt_bedpostx.html
    doc/fdt_bedpostx_parallel.html
    doc/fdt_biggest.html
    doc/fdt_display.html
    doc/fdt_dtifit.html
    doc/fdt_eddy.html
    doc/fdt_images/eddy.gif
    doc/fdt_images/eddy.tif
    doc/fdt_images/eddy2.gif
    doc/fdt_images/fdt_bedpost.gif
    doc/fdt_images/fdt_bedpost.tiff
    doc/fdt_images/fdt_bedpostx.gif
    doc/fdt_images/fdt_fa.gif
    doc/fdt_images/fdt_fa_big.png
    doc/fdt_images/fdt_gui.gif
    doc/fdt_images/fdt_gui.tiff
    doc/fdt_images/fdt_l1.gif
    doc/fdt_images/fdt_lines.gif
    doc/fdt_images/fdt_lines_subs.gif
    doc/fdt_images/fdt_probtrackx.gif
    doc/fdt_images/fdt_rgb.gif
    doc/fdt_images/fdt_rgb_subs.gif
    doc/fdt_images/fdt_seeds2targets_quant_eg.gif
    doc/fdt_images/fdt_seeds2targets_thal.gif
    doc/fdt_images/fdt_simple_tract3.gif
    doc/fdt_images/fdt_spherical_polars.gif
    doc/fdt_images/fdt_twomasks_tracts.gif
    doc/fdt_images/fdt_vecreg.gif
    doc/fdt_images/fdt_vectors.jpg
    doc/fdt_images/fdt_vectorsx.jpg
    doc/fdt_images/fdt_xfibres_axial.png
    doc/fdt_images/fdt_xfibres_coronal.png
    doc/fdt_images/fsl-bg.jpg
    doc/fdt_images/fsl-logo.jpg
    doc/fdt_images/fslview_dti.gif
    doc/fdt_images/fslview_x.gif
    doc/fdt_images/gui_intro.gif
    doc/fdt_images/gui_intro.tif
    doc/fdt_images/qboot.gif
    doc/fdt_pipeline.html
    doc/fdt_probtrackx.html
    doc/fdt_reg.html
    doc/fdt_surface.html
    doc/fdt_thresh.html
    doc/fdt_top.html
    doc/fdt_utils.html
    doc/fdt_vecreg.html
    doc/index.html
    nldtifit.cpp
    nldtifit.h
    rubix.cc
    rubix.h
    rubix_parallel.sh
    rubix_postproc.sh
    rubix_preproc.sh
    rubixoptions.cc
    rubixoptions.h
    rubixvox.cc
    rubixvox.h

+++ ptx_meshmask.cc
@@ -38,9 +38,9 @@ void meshmask()
   
   Matrix mm_to_vox(4,4);
   if(opts.meshspace.value()=="freesurfer"){
-    mm_to_vox << -1 << 0 << 0 <<  (seeds.xsize())/2
-	      <<  0 << 0 << -1 << (seeds.zsize())/2
-	      <<  0 << 1 << 0 <<  (seeds.ysize())/2
+    mm_to_vox << -1 << 0 << 0 <<  (int)seeds.xsize()/2
+	      <<  0 << 0 << -1 << (int)seeds.zsize()/2
+	      <<  0 << 1 << 0 <<  (int)seeds.ysize()/2
 	      <<  0 << 0 << 0 << 1;
   }
   else if(opts.meshspace.value()=="caret"){
